refactor: reorganize Go source tree
Closes https://github.com/scop/wrun/issues/41
